Canvas: Check size when needed.
Issue:
  If renewAllSections is called when canvas is invisible, its size is checked, and an invisible element's size seems 0.
  That 0, since it is read, is applied to the canvas element.

Usage scenario:
When Calc is hosted in an iframe and that iframe is hidden via css display: none, showing it again by removing that style results in a blank sheet. Resizing the window 'fixes' the issue.
-> If after hiding the canvas, a user triggers resize somehow, the canvas's size is set to zero (because at the time, it is not visible and its size is fetched as zero). A user can
trigger that resize event via adding / removing rows to a Calc sheet. That action in return, triggers a new sheet geometry, which in return calls resize, because only a few rows may be visible after the user action.
And we need to check that scenario. There may be other actions that may trigger this unwanted result.

Fix:
  When the element is made visible again, canvas size is still 0 (we set it manually).
  To avoid this, we check if canvas size is 0, if it is, we check the document container's size.

Adapt patch to LibreOffice Online.

Signed-off-by: Gökay Şatır <gokaysatir@gmail.com>
Change-Id: I06e3f20dd92c440d3e2f43554f9cdc34b0922b01