UI: enter edit mode in write protected file with edit button (mobile)
by default documents are opened in read only mode in mobile
to prevent accidental editing.
This read only mode is different from when file is actually
set with permission read-only(in desktop Edit -> Edit mode)

Currently online does not have any button equivalent to the desktop
to enter the edit mode, neither in desktop browser nor in mobile.

Without this patch if file is read only and user presses the pencil button,
UI changes to edit mode but file still can not be edited because of
restricted permissions.

Adapt patch to LibreOffice Online.

Signed-off-by: Pranam Lashkari <lpranam@collabora.com>
Change-Id: Id2f9b787e4a6f6456c4d192987d2650a3b2b27d0