Impress: fix preview rendering when switching to a page having different size
We assume all the pages in document to have same size. So when multiple
previews are updated at once, they disregard individual page sizes which
causes the previews to look bad.

`_invalidateCurrentPart` limits the invalidation to only current part in
case of switching to a page having different size or when only the
current page is resized.

Though there are still some cases where we invalidate tiles of multiple parts
which also updates multiple previews. That's why I'm storing individual
page sizes in `_partDimensions`

Adapt patch to LibreOffice Online.

Signed-off-by: Mohit Marathe <mohit.marathe@collabora.com>
Change-Id: Ic69da1dda0b7d43c847ca06de0ceba57ed56ef0a