Change the color of the calc headers when in sheet view
To make it more obvious that the user they are using a sheet view,
change the color of row / column headers (like in MSO) to the dark
color (or light color if in dark mode). If the current sheet view
is unsynced, then the color of the headers will change to a dark
(or light) red tone.

Adapt patch to LibreOffice Online.

Signed-off-by: Tomaž Vajngerl <tomaz.vajngerl@collabora.co.uk>
Change-Id: I677a916e90da0562c6063cc8dd7db3ef8d0ddd86