Unit test to reproduce hang when launching a new document.
This is a particularly interesting race. You need to have two live
sessions being shut-down concurrently, so each thinks the other's
in-shutdown state is not enough to send a final close, while
meanwhile a new session has been created, but not yet loaded (ie.
it hasn't created a new view.

Change-Id: I55f36f4745fc04d4b33927cd9782e58cc4615f18
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>