wsd: test: add LOK_CHECK family that doesn't break on failure
There are non-fatal checks that we might
want to do during a test, where we do
want to exercise some code, for which
the output may not be critical.

A good example is helpers used for logging
that contain timestamps that aren't
precise.

Another use-case is when we are adding
new assertions in bulk and want
to fix the expected values quickly.
With the check variant, we get to
see the output of all cases and fix
them in one go, instead of one at a time
with the edit-compile-run roundtrip.

For such cases, we want to check the
validity, so we log the result, but
don't break when the check fails.
That is, these checks are informational,
for debugging purposes only, not validating
functionality.

Change-Id: Ic4255ae119e1ec6ba5b6fa5540dc4fde60b0400d
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>