wsd: better handle view loading while destroying
There is a race when a new view is loading
while the last view is disconnecting. In
such a case, we fail to either load the view
or send back an error (such as docunloading),
to allow the client to retry later.
Instead, we are stuck, until the user reloads.

This patch handles these cases and adds a
unit-tests to verify proper handling of
such cases.

Change-Id: I660836c44092061885dbd7fd735dcef1d89da73f
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>