subclass ViewLayout with ViewLayoutWrite to make comments much more visible
here we introduced changes to improve comment visibility in writer,
previously the document was centered and whenever width/2 became
less than the comment width (which is fixed), the comments were
collapsed.

this meant if the users wanted to see the comments, they should zoom
out the document to have more space on the side margins. there was
a lot of margin space on the left side of the document which was not
utilized.

hence the code in this patch checks on zoom/resize events whether we
have enough space for comments or not and when we don't it checks if
moving (via scroll) the document to the left would make enough space
on the right side. we keep doing this till we can't get any more
space from the left side, and then we collapse the comments.

Adapt patch to LibreOffice Online.

Signed-off-by: Sahil Gautam <sahil.gautam@collabora.com>
Change-Id: I89e93518cc269e4dd9d19c6ca4e1eb6cc3ac00d8