fix(mobile): unify search icons
On computers, we have a search icon which is used for the buttons to let
us search within documents. On mobile, we have a different search icon
which is used in the quick actions bar. Unfortunately, the mobile icon
doesn't fit with the rest of the icons at all:

- It points the other way than you would expect (tilted right rather
  than tilted left)
- It is pure blue rather than the same "black with colored highlights"
  style as everything else
- It's really thin, particularly the handle which is the same width as
  the outline of the lens

This looks bad when placed next to icons that do fit, let's swap it out
so it also fits

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I8326abe38afc92460781b9c84b8499976a6a6964