feat(mobile): improve search UI layout
Before, the search UI had lots of components that (1) were not the same
height, (2) were not aligned with each other and (3) flowed off the
screen in places. I've messed with the CSS so that everything looks
much nicer...

Additionally, it makes sense to group the search bar with the thing that
clears it, so let's move them so they are next to each other

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I7aebd9b31729041d2396dbd9ce7b54826a6a6964