Fix the hyperlink popup glitch
invalidatecursor call is closing and opening hyperlink popup without
any condition. Just checking if link is exist. That causes a flicker.

In this patch we are checking if the popup is open and cursor position changed
or hyperlink is changed. We should control them before close and open action.

Adapt patch to LibreOffice Online.

Signed-off-by: Gülşah Köse <gulsah.kose@collabora.com>
Change-Id: Iec73a5ea8d75871cdd3c87d2610213bbc2ac0f97