cool#13279: SetOptimalColumnWidthDirect -> SetOptimalColumnWidth
... and fix the parameter name. For some reason, "aExtraHeight"
was sent with the command. First, SetOptimalColumnWidthDirect
does not take arguments. And next, SetOptimalColumnWidth takes
"aExtraWidth", not "aExtraHeight". The value of 200 (mm/100) is
the default value, used with SetOptimalColumnWidthDirect (see
FID_COL_OPT_DIRECT in core/sc/source/ui/view/cellsh3.cxx).

The use of SetOptimalColumnWidth is a pre-requisite to passing
also the column, taken from the clicked column header.

Signed-off-by: Mike Kaganski <mike.kaganski@collabora.com>
Change-Id: Iba307aeb3a587c41fc064d8077b0ef35039685ab