Remove second, redundant call of UnitBase::uninit
...which is already unconditionally called in COOLWSD::innerMain.

As Ash said at
<https://matrix.to/#/!ZLwFhFGGuRuiYyBrqf:libera.chat/$1Uk4s8GY631h3GcINW_2WBS8Pgq6P64x6nj9CcSjLnI?via=ericeira.cloud&via=t2bot.io&via=matrix.org>:
"There was a scenario (that I can't remember now) that made this necessary.
Perhaps an early failure? An exception? But I'm sure it was intentional because
it was needed. Perhaps we can simplify it and return it only when needed (this
time with a proper comment) ?"

Adapt patch to LibreOffice Online.

Signed-off-by: Stephan Bergmann <stephan.bergmann@collabora.com>
Change-Id: I9fc3c1e4c6f34003d1c8952453438425f16b72b5