Make copyVersionInfoToClipboard() more robust
For example in case of mobile apps some info is missing
from the about box, such as os-info, coolwsd-id and
wopi-host-id. Instead of "Uncaught TypeError: Cannot
read properties of null (reading 'innerText')" we
silently ignore and do not try to copy missing info
to the clipboard.

Adapt patch to LibreOffice Online.

Signed-off-by: Andras Timar <andras.timar@collabora.com>
Change-Id: Ic1d05781ddf60491499a1e479d442c190c965be0