FakeSocket - add a 'Queued' write wrapper.
In many cases we want an instant 'buffer this message' write, that
does a full-write without blocking.

fakeSocketWrite() already does this, and POLLOUT is flagged on
every FakeSocket that is not shutdown; nor is there any error handling
at any of the call sites. cf. FakeSocket.cpp /checkForPoll/

So wrap this in a #define wrapper for now in case we need to
re-introduce a synchronous poll for OUT space.

Ultimately the data stream from the browser is typically of a
manageable size, such that we can buffer it all without flow
control, and trust to fast event compression on the read side.

Finally remove the extremely unhelpful idea of launching a thread
to do each write before it catches on; no need for that raciness,
inefficiency and risk of event re-ordering.

Adapt patch to LibreOffice Online.

Signed-off-by: Michael Meeks <michael.meeks@collabora.com>
Change-Id: I6cbf0c70730cc7cbb644ecf959ddffb73a3c605d