use a checkbox for 'showruler' widget on writer's notebookbar
previously we used a helper which listened for the 'rulerchanged'
event and added/removed the selected class from the button which
modified it's css to show it's state.

now for the checkbox, we set the ruler's state in 'stateChangeHandler'
and fire `commandstatechanged` event, which the checkbox widget listens
for and it sets it's state to the value set in `stateChangeHandler` map.

Signed-off-by: Sahil Gautam <sahil.gautam@collabora.com>
Change-Id: I3eb93b402493fc89f853f3473ab47e6e6a6a5fd9