fix(browserstack): Specify provisioning profile
To export an archive from xcodebuild, you need a provisioning profile.

In my testing environment, I was logged into xcode - so it generated
one for me and picked it automatically. In CI, xcode won't pick up on a
certificate it didn't generate automatically, so we need to specify one
here manually...

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I077867212040528bbb12d3f1d86078c76a6a6964